Revert "Add memory log level and memory logs correspondingly (#6037)"
This reverts commit 70efb12f5f75c7b1642542e7c3ce330c472ea038.
