+++ runtime/include/tt/runtime/detail/distributed/distributed.h
@@ -12,8 +12,6 @@ namespace tt::runtime::distributed {
 void launchDistributedRuntime(const DistributedOptions &options = {});
 void shutdownDistributedRuntime();
 
-void setMemoryLogLevel(const MemoryLogLevel &logLevel);
-
 SystemDesc getCurrentSystemDesc(
     std::optional<::tt::runtime::DispatchCoreType> dispatchCoreType =
         std::nullopt,
