+++ runtime/include/tt/runtime/detail/distributed/flatbuffer/command.fbs
@@ -13,11 +13,6 @@ table ConfigureRuntimeContextCommand {
   mlir_home: string;
   metal_home: string;
   current_device_runtime: tt.runtime.flatbuffer.DeviceRuntime;
-  memory_log_level: tt.runtime.flatbuffer.MemoryLogLevel;
-}
-
-table SetMemoryLogLevelCommand {
-  memory_log_level: tt.runtime.flatbuffer.MemoryLogLevel;
 }
 
 table GetSystemDescCommand {
@@ -147,7 +142,6 @@ table ShutdownCommand {
 
 union CommandType {
   ConfigureRuntimeContextCommand,
-  SetMemoryLogLevelCommand,
   GetSystemDescCommand,
   SetFabricConfigCommand,
   GetNumAvailableDevicesCommand,
