+++ runtime/include/tt/runtime/detail/distributed/worker/response_factory.h
@@ -18,10 +18,6 @@ public:
                                  const std::string &errorMessage);
 
   static void
-  buildSetMemoryLogLevelResponse(::flatbuffers::FlatBufferBuilder &fbb,
-                                 uint64_t commandId);
-
-  static void
   buildConfigureRuntimeContextResponse(::flatbuffers::FlatBufferBuilder &fbb,
                                        uint64_t commandId);
