+++ runtime/lib/distributed/controller/controller.cpp
@@ -10,7 +10,6 @@
 #include "tt/runtime/detail/distributed/flatbuffer/flatbuffer.h"
 #include "tt/runtime/detail/distributed/utils/utils.h"
 #include "tt/runtime/runtime.h"
-#include "tt/runtime/utils.h"
 namespace tt::runtime::distributed::controller {
 
 namespace fb = ::tt::runtime::distributed::flatbuffer;
@@ -227,18 +226,6 @@ void Controller::setReadTimeout(const std::chrono::seconds &timeout) {
 
 void Controller::setWorkerShutdownTimeout(const std::chrono::seconds &timeout) {
   workerShutdownTimeout_ = timeout;
-}
-
-void Controller::setMemoryLogLevel(const MemoryLogLevel &logLevel) {
-
-  auto commandBuilder = std::make_unique<::flatbuffers::FlatBufferBuilder>();
-
-  uint64_t commandId =
-      CommandFactory::buildSetMemoryLogLevelCommand(*commandBuilder, logLevel);
-
-  pushToCommandAndResponseQueues(commandId,
-                                 fb::CommandType::SetMemoryLogLevelCommand,
-                                 std::move(commandBuilder));
 }
 
 SystemDesc Controller::getCurrentSystemDesc(
@@ -808,8 +795,7 @@ void Controller::configureRuntimeContext() {
   auto commandId = CommandFactory::buildConfigureRuntimeContextCommand(
       *commandBuilder, ::tt::runtime::RuntimeContext::instance().getMlirHome(),
       ::tt::runtime::RuntimeContext::instance().getMetalHome(),
-      ::tt::runtime::RuntimeContext::instance().getCurrentDeviceRuntime(),
-      ::tt::runtime::RuntimeContext::instance().getMemoryLogLevel());
+      ::tt::runtime::RuntimeContext::instance().getCurrentDeviceRuntime());
 
   auto awaitingPromise = std::make_unique<std::promise<void>>();
   std::future<void> awaitingFuture = awaitingPromise->get_future();
@@ -861,18 +847,6 @@ void Controller::handleConfigureRuntimeContextResponse(
   DEBUG_ASSERT(awaitingPromise, "Awaiting promise must be populated");
 
   awaitingPromise->set_value();
-}
-
-void Controller::handleSetMemoryLogLevelResponse(
-    const std::vector<SizedBuffer> &responseBuffers,
-    std::unique_ptr<AwaitingResponseQueueEntry> awaitingResponse) {
-
-  debug::checkResponsesIdentical(responseBuffers);
-
-  debug::checkResponseTypes(responseBuffers,
-                            fb::ResponseType::SetMemoryLogLevelResponse);
-
-  debug::assertNoAwaitingState(*awaitingResponse, "SetMemoryLogLevel");
 }
 
 void Controller::handleGetSystemDescResponse(
@@ -1303,10 +1277,6 @@ void Controller::handleResponse(
   case fb::CommandType::ConfigureRuntimeContextCommand: {
     return handleConfigureRuntimeContextResponse(responseBuffers,
                                                  std::move(awaitingResponse));
-  }
-  case fb::CommandType::SetMemoryLogLevelCommand: {
-    return handleSetMemoryLogLevelResponse(responseBuffers,
-                                           std::move(awaitingResponse));
   }
   case fb::CommandType::GetSystemDescCommand: {
     return handleGetSystemDescResponse(responseBuffers,
