+++ runtime/lib/runtime.cpp
@@ -126,23 +126,6 @@ void setMetalHome(std::string_view metalHome) {
   LOG_FATAL("runtime is not enabled");
 }
 
-void setMemoryLogLevel(const MemoryLogLevel &logLevel) {
-  using RetType = void;
-  return DISPATCH_TO_CURRENT_RUNTIME(
-      RetType,
-      [&]() -> RetType {
-        return ::tt::runtime::RuntimeContext::instance().setMemoryLogLevel(
-            logLevel);
-      },
-      [&]() -> RetType {
-        return ::tt::runtime::RuntimeContext::instance().setMemoryLogLevel(
-            logLevel);
-      },
-      [&]() -> RetType {
-        return ::tt::runtime::distributed::setMemoryLogLevel(logLevel);
-      });
-}
-
 std::vector<DeviceRuntime> getAvailableDeviceRuntimes() {
   std::vector<DeviceRuntime> runtimes;
 #if defined(TT_RUNTIME_ENABLE_TTNN) && (TT_RUNTIME_ENABLE_TTNN == 1)
