+++ runtime/lib/ttnn/runtime.cpp
@@ -38,6 +38,28 @@ namespace tt::runtime::ttnn {
 
 using ::tt::runtime::DeviceRuntime;
 
+namespace debug {
+static void logDeviceMemoryState(const Device &device,
+                                 std::string_view prefix) {
+#if defined(TT_RUNTIME_DEBUG) && TT_RUNTIME_DEBUG == 1
+  ::tt::runtime::debug::logMemoryState(getMemoryView(device), prefix);
+#endif
+}
+} // namespace debug
+
+static tt::runtime::MemoryView
+createMemoryView(const tt::tt_metal::detail::MemoryView &memoryView) {
+  return tt::runtime::MemoryView{
+      .numBanks = memoryView.num_banks,
+      .totalBytesPerBank = memoryView.total_bytes_per_bank,
+      .totalBytesAllocatedPerBank = memoryView.total_bytes_allocated_per_bank,
+      .totalBytesFreePerBank = memoryView.total_bytes_free_per_bank,
+      .largestContiguousBytesFreePerBank =
+          memoryView.largest_contiguous_bytes_free_per_bank,
+      .blockTable = memoryView.block_table,
+  };
+}
+
 template <typename T>
 static ::ttnn::Tensor createBorrowedTTNNTensor(void *rawData,
                                                const ::ttnn::Shape &shape) {
@@ -666,7 +688,29 @@ void readDeviceProfilerResults(Device deviceHandle) {
 
 std::unordered_map<tt::runtime::MemoryBufferType, tt::runtime::MemoryView>
 getMemoryView(Device deviceHandle) {
-  return utils::getMemoryView(deviceHandle);
+  std::unordered_map<tt::runtime::MemoryBufferType, tt::runtime::MemoryView>
+      memoryMap;
+  ::ttnn::MeshDevice &meshDevice =
+      deviceHandle.as<::ttnn::MeshDevice>(DeviceRuntime::TTNN);
+
+  auto dramMemoryView = ::tt::tt_metal::detail::GetMemoryView(
+      &meshDevice, ::ttnn::BufferType::DRAM);
+  auto l1MemoryView = ::tt::tt_metal::detail::GetMemoryView(
+      &meshDevice, ::ttnn::BufferType::L1);
+  auto l1SmallMemoryView = ::tt::tt_metal::detail::GetMemoryView(
+      &meshDevice, ::ttnn::BufferType::L1_SMALL);
+  auto traceMemoryView = ::tt::tt_metal::detail::GetMemoryView(
+      &meshDevice, ::ttnn::BufferType::TRACE);
+
+  memoryMap[tt::runtime::MemoryBufferType::DRAM] =
+      createMemoryView(dramMemoryView);
+  memoryMap[tt::runtime::MemoryBufferType::L1] = createMemoryView(l1MemoryView);
+  memoryMap[tt::runtime::MemoryBufferType::L1_SMALL] =
+      createMemoryView(l1SmallMemoryView);
+  memoryMap[tt::runtime::MemoryBufferType::TRACE] =
+      createMemoryView(traceMemoryView);
+
+  return memoryMap;
 }
 
 void setFabricConfig(tt::runtime::FabricConfig config) {
@@ -1699,10 +1743,8 @@ std::vector<::tt::runtime::Tensor>
 submit(Device deviceHandle, Binary executableHandle, std::uint32_t programIndex,
        std::vector<::tt::runtime::Tensor> &inputs) {
 
-  ::tt::runtime::utils::logMemoryStateIfNeeded(
-      utils::getMemoryView(deviceHandle),
-      ::tt::runtime::MemoryLogLevel::Program,
-      "Device memory state before submit");
+  debug::logDeviceMemoryState(deviceHandle,
+                              "Device memory state before submit");
 
   std::unique_ptr<ProgramExecutor> executor = std::make_unique<ProgramExecutor>(
       deviceHandle, executableHandle, programIndex, inputs);
@@ -1713,10 +1755,7 @@ submit(Device deviceHandle, Binary executableHandle, std::uint32_t programIndex,
 
   executor.reset();
 
-  ::tt::runtime::utils::logMemoryStateIfNeeded(
-      utils::getMemoryView(deviceHandle),
-      ::tt::runtime::MemoryLogLevel::Program,
-      "Device memory state after submit");
+  debug::logDeviceMemoryState(deviceHandle, "Device memory state after submit");
 
   return outputTensors;
 }
