+++ runtime/python/runtime/runtime.cpp
@@ -292,18 +292,10 @@ void registerRuntimeBindings(nb::module_ &m) {
       .value("WORMHOLE_B0", ::tt::target::Arch::Wormhole_b0)
       .value("BLACKHOLE", ::tt::target::Arch::Blackhole);
 
-  nb::enum_<::tt::runtime::MemoryLogLevel>(m, "MemoryLogLevel", nb::is_flag())
-      .value("NONE", ::tt::runtime::MemoryLogLevel::NONE)
-      .value("PROGRAM", ::tt::runtime::MemoryLogLevel::Program)
-      .value("OPERATION", ::tt::runtime::MemoryLogLevel::Operation)
-      .value("ANY", ::tt::runtime::MemoryLogLevel::ANY);
-
   m.def("set_mlir_home", &tt::runtime::setMlirHome, nb::arg("mlir_home"),
         "Set the MLIR home directory");
   m.def("set_metal_home", &tt::runtime::setMetalHome, nb::arg("metal_home"),
         "Set the Metal home directory");
-  m.def("set_memory_log_level", &tt::runtime::setMemoryLogLevel,
-        nb::arg("log_level"), "Set the memory log level");
   m.def("get_current_device_runtime", &tt::runtime::getCurrentDeviceRuntime,
         "Get the backend device runtime type");
   m.def("get_available_device_runtimes",
@@ -650,5 +642,8 @@ void registerRuntimeBindings(nb::module_ &m) {
 
   m.def("unregister_hooks",
         []() { ::tt::runtime::debug::Hooks::get().unregisterHooks(); });
+
+  m.def("log_memory_state", &::tt::runtime::debug::logMemoryState,
+        nb::arg("memory_state"), nb::arg("prefix") = "");
 }
 } // namespace tt::runtime::python
